requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-dotenv==1.1.0

# Telegram integration
Pyrogram==2.0.106
TgCrypto==1.2.5

# API models and validation
pydantic==2.11.5
pydantic_core==2.33.2

# HTTP client
requests==2.32.4
httpx==0.28.1

# Async support
anyio==4.9.0
sniffio==1.3.1

# Security
PyJWT==2.10.1
cryptography==44.0.0

# Type hints support (for Python 3.12 compatibility)
typing_extensions==4.14.0
annotated-types==0.7.0

# Additional utilities
python-multipart==0.0.20
